requirements.txt
torch>=2.1
torchvision>=0.16
torchaudio>=2.1
timm>=0.9.12
numpy>=1.23
pandas>=2.0
scikit-learn>=1.3
Pillow>=10.0
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.7
